Refactored Tests;
