+++ Tests/BootCamp.Chapter.Tests/CityTest.cs
@@ -2,6 +2,7 @@ using System;
 using Xunit;
 using BootCamp.Chapter.Models;
 using BootCamp.Chapter.Exceptions;
+using BootCamp.Chapter.Tests.Public;
 
 namespace BootCamp.Chapter.Tests
 {
@@ -12,7 +13,7 @@ namespace BootCamp.Chapter.Tests
         [InlineData(1, 51)]
         public void City_Given_BadBuildings_Throws_BuildingException(int citySize, int buildingHeight)
         {
-            int[][] buildings = BuildBuildingsArray(citySize, buildingHeight);
+            int[][] buildings = Builders.BuildBuildingsArray(citySize, buildingHeight);
             City city;
 
             Action action = () => city = new City(buildings);
@@ -20,24 +21,7 @@ namespace BootCamp.Chapter.Tests
             Assert.Throws<BuildingException>(action);
         }
 
-        [Theory]
-        [InlineData(true, 4, 1, 50)]
-        [InlineData(false, 4, 1, 50)]
-        public void SkyLine_Given_Buildings_Creates_View(bool isTopView, int citySize, int buildingHeight, int offNumber)
-        {
-            //Arrange
-            int[][] buildings = BuildBuildingsArray(citySize, buildingHeight);
-            buildings[0][0] = offNumber;
-            int[] correctView = new int[4] { offNumber, buildingHeight , buildingHeight, buildingHeight };
-
-            //Act
-            SkyLine skyLine = new SkyLine(buildings);
-            int[] view = isTopView ? skyLine.TopView : skyLine.SideView;
-
-            //Assert
-            Assert.Equal(correctView, view);
-        }
-
+        /*
         [Theory]
         [InlineData(0, 3, 1, 5)]
         [InlineData(1, 3, 1, 5)]
@@ -56,7 +40,7 @@ namespace BootCamp.Chapter.Tests
             //Assert
             Assert.Equal(correctMaxBuildingHeight, MaxHeightBuildings);
         }
-
+        
         private int[][] GetMaxBuildingsArrArr(int citySize, int buildingHeight, int offNumber, int offPlace)
         {
             int[][] buildings = new int[citySize][];
@@ -72,20 +56,6 @@ namespace BootCamp.Chapter.Tests
             }
             return buildings;
         }
-
-        private int[][] BuildBuildingsArray(int citySize, int buildingHeight)
-        {
-            int[][] buildings = new int[citySize][];
-            for (int i = 0; i < citySize; i++)
-            {
-                buildings[i] = new int[citySize];
-                for (int k = 0; k < citySize; k++)
-                {
-                    buildings[i][k] = buildingHeight;
-                }
-            }
-
-            return buildings;
-        }
+        */
     }
 }
